// angular-app/src/app/services/chat-services/models/notification-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NotificationDto {
  chatId?: string;
  chatName?: string;
  content?: string;
  id?: number;
  media?: string;
  messageType?: 'TEXT' | 'IMAGE' | 'VIDEO' | 'AUDIO';
  read?: boolean;
  receiverId?: string;
  relatedEntityId?: number;
  relatedEntityType?: string;
  senderId?: string;
  type?: 'SEEN' | 'MESSAGE' | 'IMAGE' | 'AUDIO' | 'VIDEO';
}