// angular-app/src/app/services/group-services/models/notification-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface NotificationDto {
  groupId?: string;
  groupName?: string;
  id?: number;
  message?: string;
  read?: boolean;
  relatedEntityId?: number;
  relatedEntityType?: string;
  timestamp?: string;
  type?: string;
  userId?: string;
  userName?: string;
}